// src/app/admin/models/enroll.ts
export interface Enroll {
    enrollDate: string;
    id: string;
    courseId: number;
    isApproved: boolean;
    isCompleted: boolean;
    isPaid: boolean;
    isRefunded: boolean;
    isFree: boolean;
    isTrial: boolean;
    isActive: boolean;
    isReject: boolean;
    instructorDetailsId: number;
    courseAssignId: number;
}